Create Cal Form. WITH ForEach Calc.

+++ src/pratic10/CalcArea/Form.java
@@ -1,5 +1,5 @@
 package pratic10.CalcArea;
 
 public interface Form {
-    int calcArea();
+    double calcArea();
 }
